+++ app/src/main/AndroidManifest.xml
@@ -14,6 +14,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
+
         android:theme="@style/Theme.WellAging"
         tools:targetApi="31">
         <activity
